refactored the views and changed icon behavior
