+++ src/views/defaults/index.ts
@@ -1,0 +1,15 @@
+/* tslint:disable:ordered-imports */
+import ViewTemplates from '@lit-any/lit-any/views'
+import './hydrofoil-shell'
+import './alcaeus-resource-viewer'
+import './resource-link'
+import './zoom-to-blank-icon'
+import {html} from 'lit-html'
+
+ViewTemplates.default.when
+  .valueMatches((v) => typeof v === 'string')
+  .renders((v) => v)
+
+ViewTemplates.default.when
+  .valueMatches((v) => Array.isArray(v))
+  .renders((v: any[], render) => v.map((el) => html`${render(el)} `))
